schema is moved to separate file
