+++ controllers/contacts.js
@@ -2,14 +2,6 @@ const functions = require("../models/contacts");
 
 const { HttpError, ctrlWrapper } = require("../helpers");
 
-const Joi = require("joi");
-
-const addSchema = Joi.object({
-	name: Joi.string().min(3).required(),
-	email: Joi.string().required(),
-	phone: Joi.string().min(7).required(),
-});
-
 const listContacts = async (req, res) => {
 	const result = await functions.listContacts();
 	res.json(result);
@@ -25,10 +17,6 @@ const getById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-	const { error } = addSchema.validate(req.body);
-	if (error) {
-		throw HttpError(400, error.message);
-	}
 	const result = await functions.addContact(req.body);
 	res.status(201).json(result);
 };
@@ -45,10 +33,6 @@ const deleteContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
-	const { error } = addSchema.validate(req.body);
-	if (error) {
-		throw HttpError(400, error.message);
-	}
 	const { id } = req.params;
 	const result = await functions.updateContact(id, req.body);
 	if (!result) {
