+++ routes/api/contacts.js
@@ -2,16 +2,20 @@ const express = require("express");
 
 const router = express.Router();
 
-const ctrl = require("../../controllers/contacts")
+const ctrl = require("../../controllers/contacts");
+
+const { validateBody } = require("../../middlewares");
+
+const schemas = require("../../schemas/contacts");
 
 router.get("/", ctrl.listContacts);
 
-router.get("/:id", ctrl.getById );
+router.get("/:id", ctrl.getById);
 
-router.post("/", ctrl.addContact);
+router.post("/", validateBody(schemas.addSchema), ctrl.addContact);
 
 router.delete("/:id", ctrl.deleteContact);
 
-router.put("/:id", ctrl.updateContact);
+router.put("/:id", validateBody(schemas.addSchema), ctrl.updateContact);
 
 module.exports = router;
