+++ schemas/contacts.js
@@ -1,0 +1,11 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+	name: Joi.string().min(3).required(),
+	email: Joi.string().required(),
+	phone: Joi.string().min(7).required(),
+});
+
+module.exports = {
+	addSchema,
+};
